requirements.txt
ffmpeg-python==0.2.0
future==0.18.2
python-ffmpeg-video-streaming>=0.1

urllib3~=1.26.4
scte~=1.1.4
mpegdash~=0.2.0
m3u8~=0.8.0